// types.ts
export interface UserProfile {
    id: string;
    email: string;
    role: 'master' | 'student';
    name: string;
    avatarUrl: string;
    academyId: string;
    studentId?: string;
    password?: string;
    emailConfirmed?: boolean;
    paymentMethods?: { id: string; brand: string; last4: string; expMonth: number; expYear: number }[];
}